Move RepositoryComponent and graph components into its own file
